+++ nutree/typed_tree.py
@@ -29,9 +29,9 @@ from nutree.common import (
     CycleDetectedError,
     DataIdType,
     DeserializeMapperType,
+    DotMapperCallbackType,
     IterMethod,
     KeyMapType,
-    MapperCallbackType,
     PredicateCallbackType,
     SerializeMapperType,
     UniqueConstraintError,
@@ -533,8 +533,8 @@ class TypedNode(Node[TData]):
         graph_attrs: dict | None = None,
         node_attrs: dict | None = None,
         edge_attrs: dict | None = None,
-        node_mapper: MapperCallbackType | None = None,
-        edge_mapper: MapperCallbackType | None = None,
+        node_mapper: DotMapperCallbackType | None = None,
+        edge_mapper: DotMapperCallbackType | None = None,
     ) -> Iterator[str]:
         """Generate a DOT formatted graph representation.