+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Dropping Python versions that reached EOL will only cause a minor version bump.
 
 - New method `node.parent_iterator()`.
 - New option `Tree(..., check_dag=True)` to enforce/relax DAG compliance (default: true).
+- Add `tree.map(fn)`
+- BREAKING: Rename `MapperCallbackType` to `DotMapperCallbackType`
 - Drop Python 3.9 (EOL 2025-10)
 - Support Python 3.14, 3.15
 - Use [uv](https://docs.astral.sh/uv/)